pyproject.toml
[project]
name = "xai-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "captum",
    "datasets>=4.5.0",
    "ipykernel>=7.2.0",
    "medmnist>=3.0.2",
    "open-clip-torch>=3.3.0",
    "pillow>=12.1.0",
    "quantus[torch,transformers]",
    "seaborn>=0.13.2",
    "timm>=1.0.25",
    "torchvision>=0.25.0",
]

[tool.uv.sources]
quantus = { path = "./modules/Quantus", editable = true }
captum = { path = "./modules/captum", editable = true }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
